Create MainActivity.java
This activity creates a map with a marker that checks the user's location.  The marker's text is "other" if no GPS data is found, "ready" if the user is near the point or "waiting" if the player is too far away.

Accuracy is based on the amount of leeway around the marker (0.0005 is aprx 15 metres).

It should pick the best provider available, but it just picks the GPS provider instead.